backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
dj-database-url==2.2.0
gunicorn==22.0.0
whitenoise==6.7.0
python-dotenv==1.0.1
